🧱 Make an abstract class for JsonSourceCopier
This is pretty hack-and-slash. I've basically pulled out anything that looks 5e specific into an abstract method, and then overrode that in the 5e implementation.

The end goal of this is to make a Pf2e implementation.

Confirmed with diff that this doesn't change the output of any 5e notes.
